+++ staramr/blast/pointfinder/PointfinderDatabaseInfo.py
@@ -1,18 +1,27 @@
 import pandas as pd
+import logging
+
+from os import path
 
 """
 A Class storing information about the specific PointFinder database.
 """
 
+logger = logging.getLogger('PointfinderDatabaseInfo')
+
 
 class PointfinderDatabaseInfo:
 
-    def __init__(self, database_info_dataframe):
+    def __init__(self, database_info_dataframe, file=None):
         """
         Creates a new PointfinderDatabaseInfo.
         :param database_info_dataframe: A pd.DataFrame containing the information in PointFinder.
+        :param file: The file where the pointfinder database info originates from.
         """
         self._pointfinder_info = database_info_dataframe
+        self._file = file
+
+        self._resistance_table_hacks(self._pointfinder_info)
 
     @classmethod
     def from_file(cls, file):
@@ -22,8 +31,8 @@ class PointfinderDatabaseInfo:
         :param file: The file containing drug resistance mutations.
         :return: A new PointfinderDatabaseInfo.
         """
-        pointfinder_info = pd.read_table(file, index_col=False)
-        return cls(pointfinder_info)
+        pointfinder_info = pd.read_csv(file, sep='\t', index_col=False)
+        return cls(pointfinder_info, file)
 
     @classmethod
     def from_pandas_table(cls, database_info_dataframe):
@@ -33,6 +42,18 @@ class PointfinderDatabaseInfo:
         :return: A new PointfinderDatabaseInfo.
         """
         return cls(database_info_dataframe)
+
+    def _resistance_table_hacks(self, table):
+        """
+        A function implementing some hacks to try and fix mismatched strings in the pointfinder databases.
+        These should be removed when the underlying database is corrected.
+        :param table: The pointfinder resistance table to fix.
+        :return: None, but modifies the passed table in place.
+        """
+        if self._file and 'salmonella' in str(self._file) and path.exists(
+                path.join(path.dirname(self._file), '16S_rrsD.fsa')):
+            logger.debug("Replacing [16S] with [16S_rrsD] for pointfinder organism [salmonella]")
+            table[['#Gene_ID']] = table[['#Gene_ID']].replace('16S', '16S_rrsD')
 
     def _get_resistance_codon_match(self, gene, codon_mutation):
         table = self._pointfinder_info